+++ src/confidential_key_derivation/test.rs
@@ -5,24 +5,24 @@ use crate::Participant;
 
 #[test]
 fn test_keygen() {
-    let participants = vec![
+    let participants = [
         Participant::from(31u32),
         Participant::from(1u32),
         Participant::from(2u32),
     ];
     let threshold = 2;
-    crate::dkg::test::test_keygen::<C>(participants, threshold);
+    crate::dkg::test::test_keygen::<C>(&participants, threshold);
 }
 
 #[test]
 fn test_refresh() {
-    let participants = vec![
+    let participants = [
         Participant::from(0u32),
         Participant::from(31u32),
         Participant::from(2u32),
     ];
     let threshold = 2;
-    crate::dkg::test::test_refresh::<C>(participants, threshold);
+    crate::dkg::test::test_refresh::<C>(&participants, threshold);
 }
 
 #[test]
@@ -30,5 +30,5 @@ fn test_reshare() {
     let participants = generate_participants(3);
     let threshold0 = 2;
     let threshold1 = 3;
-    crate::dkg::test::test_reshare::<C>(participants, threshold0, threshold1)
+    crate::dkg::test::test_reshare::<C>(&participants, threshold0, threshold1);
 }
